requirements.txt
cffi==1.17.1
cryptography==45.0.5
pycparser==2.22
pyOpenSSL==25.1.0
